Remove `View Report`
Resolves #207.

Both the Bundle list and Detail have a View Report option. However, since their
current directories differ, one will result in a 404 error. This started
occurring after upgrading to rancher/shell 3.0.7. Even if View Report is
removed, the report can still be viewed in Detail, so this PR removes it.

Signed-off-by: Masashi Honma <masashi.honma@gmail.com>